--- packages/mobile/src/components/templates/FullModalTemplate/style.module.scss
@@ -1,3 +1,0 @@
-.modal {
-  background: white;
-}
