+++ packages/mobile/src/components/templates/FullPageModalTemplate/style.module.scss
@@ -1,8 +1,7 @@
 @import '../../../styles/main.scss';
 
-.subscription {
+.full-modal {
   height: 100vh;
-  padding: 0 25px;
   background: #fff;
 
   .header {
