+++ packages/mobile/src/page/Subscription/style.module.scss
@@ -1,5 +1,9 @@
 @import '../../styles/main.scss';
 
+.modal {
+  padding: 0 25px;
+  background: #fff;
+}
 
 .guide {
   padding-top: 34px;
